Day 3-2 Java - correct datatypes

+++ 2020/java/src/main/java/site/markhenrick/adventofcode2020/Day03.java
@@ -35,8 +35,8 @@ class Day03 {
 		return character;
 	}
 
-	private int solveForDescentVector(final IntVector descentVector) {
-		return (int) Stream.iterate(
+	private long solveForDescentVector(final IntVector descentVector) {
+		return Stream.iterate(
 			IntVector.ORIGIN,
 			location -> location.x <= maxX,
 			location -> IntVector.add(location, descentVector)
